Crate gfx_text

A library for drawing text for gfx-rs graphics API. Uses freetype-rs underneath to former the font bitmap texture and collect information about face glyphs.

Examples

Basic usage:

// Initialize text renderer.
let mut text = gfx_text::new(factory).build().unwrap();

// In render loop:

// Draw some text 10 pixels down and right from the top left screen corner.
text.draw(
    "The quick brown fox jumps over the lazy dog",  // Text to draw
    [10, 10],                                       // Position
    [0.65, 0.16, 0.16, 1.0],                        // Text color
);

// Render the final batch.
text.draw_end(&mut stream);

Structs

Renderer
Text renderer.

RendererBuilder
Text renderer builder. Allows to set rendering options using builder pattern.

Enums

Error
General error type returned by the library. Wraps all other errors.

FontError
Represents possible errors which may occur during the font loading.

Functions

new
Create a new text renderer builder. Alias for RendererBuilder::new.