Crate gfx_text

A library for drawing text for gfx-rs graphics API. Uses freetype-rs underneath to former the font bitmap texture and collect information about face glyphs.

Structs

Renderer
Text renderer instance.

RendererBuilder
Text renderer builder instance.

Enums

Error
General error type returned by the library. Wraps other errors which may occur during some operations.

FontError
Represents possible errors that may occur during the font loading.

Functions

new
Create a new text renderer builder. Alias for RendererBuilder::new.