Struct gfx_text::Renderer

pub struct Renderer<R: Resources, F: Factory<R>> {
    // some fields omitted
}

Text renderer instance.

Methods

impl<R: Resources, F: Factory<R>> Renderer<R, F>

fn draw(&mut self, text: &str, pos: [i32; 2], color: [f32; 4])

Add some text to the current draw scene relative to the top left corner of the screen using pixel coords.

fn draw_at(&mut self, text: &str, pos: [f32; 3], color: [f32; 4])

Add some text to the draw scene using absolute world coords.

fn draw_end<S: Stream<R>>(&mut self, stream: &mut S) -> Result<(), Error>

End with drawing, clear internal state and return resulting batch.

fn draw_end_at<S: Stream<R>>(&mut self, stream: &mut S, proj: [[f32; 4]; 4]) -> Result<(), Error>

End with drawing using provided projection matrix.

fn get_batch<O: Output<R>>(&mut self, output: &O) -> Result<OwnedBatch<ShaderParams<R>>, Error>

End with drawing and former resulting batch.

fn get_batch_at<O: Output<R>>(&mut self, output: &O, proj: [[f32; 4]; 4]) -> Result<OwnedBatch<ShaderParams<R>>, Error>

Return batch for the given projection matrix.