Struct gfx_text::RendererBuilder

pub struct RendererBuilder<'r, R: Resources, F: Factory<R> + 'r> {
    // some fields omitted
}

Text renderer builder instance.

Methods

impl<'r, R: Resources, F: Factory<R>> RendererBuilder<'r, R, F>

fn new(factory: &'r mut F) -> Self

Create a new text renderer builder.

fn with_size(self, size: u8) -> Self

Specify custom size.

fn with_font(self, path: &'r str) -> Self

Specify custom font by path.

fn with_font_data(self, data: &'r [u8]) -> Self

Pass raw font data.

fn with_outline(self, width: u8, color: [f32; 4]) -> Self

Specify outline width and color.

fn with_buffer_size(self, size: usize) -> Self

Specify custom initial buffer size.

fn with_chars(self, chars: &'r [char]) -> Self

Make available only provided characters in font texture instead of loading all existing from the font face.

fn build(self) -> Result<Renderer<R>, Error>

Build a new text renderer instance using current settings.

fn unwrap(self) -> Renderer<R>

Just an alias for builder.build().unwrap().